Update the display name of a User and associated addresses in Core.
Currently, only when a User updates their name using the User Profile view that
their name is updated in Core. This is good for the user but prevents an Admin
from changing their display name properly since they don't have access to the
User's user-profile page even though they should be able to update their name.

This adds a signal handler that will update the display name regardless of how
it was updated. A superuser can update the name in Admin dashboard and that
will be reflected in Core.

Also, this adds support for updating the display_name in every associated
address of the User. It will override whatever the address was initially set.