Don't assume username exists on AUTH_USER model
If you want to use django-mailman3 as an `INSTALLED_APP` in
a Django project that has a User model that does not have a username
field (typically email login), the assumptions in the code base
prevent that.

This thankfully is mostly in the log messages and `__str__` methods
so it's easy to swap out the `username` for the `pk` to still
create a meaningful log message.